+++ forge/src/main/java/com/christofmeg/justenoughbreeding/config/JEBIntegration.java
@@ -145,5 +145,11 @@ public class JEBIntegration {
         if (ModList.get().isLoaded("crittersandcompanions")) {
             CrittersAndCompanions.init();
         }
+        if (ModList.get().isLoaded("ambientadditions")) {
+            AmbientAdditionsIntegration.init();
+        }
+        if (ModList.get().isLoaded("tfc")) {
+            TerraFirmaCraftIntegration.init();
+        }
     }
 }