+++ fabric/changelog.md
@@ -1,1 +1,2 @@
-Port to 1.21
+1.2.5:
+- Update version range to 1.21.1