+++ neoforge/changelog.md
@@ -1,1 +1,6 @@
-Port to 1.21
+1.3.0:
+- Ambient Additions Integration
+- Terra Firma Craft Integration
+
+1.2.5:
+- Update version range to 1.21.1